Исправление ошибки зацикливания модулей
